requirements.txt
Flask
Werkzeug
flask-login
Pillow